remove styles
